fix: transform shorthand properties in directive transform
