+++ src/nuxt/module.ts
@@ -6,10 +6,9 @@ import {
   createResolver,
   defineNuxtModule,
 } from '@nuxt/kit'
-import { NodeTypes, createSimpleExpression } from '@vue/compiler-core'
-import { stringifyStyle } from '@vue/shared'
-
 import type { NuxtModule } from '@nuxt/schema'
+import motionDirectiveTransform from '../directive/transform'
+
 import type { ModuleOptions as MotionModuleOpts } from '../types'
 
 export interface ModuleOptions extends MotionModuleOpts<string> {}
@@ -49,72 +48,17 @@ export default defineNuxtModule<ModuleOptions>({
     })
 
     nuxt.options.vue.compilerOptions.directiveTransforms ??= {}
-    nuxt.options.vue.compilerOptions.directiveTransforms.motion = (
-      dir,
-      node,
-      _context,
-    ) => {
-      for (const property of node.props) {
-        // only handle :initial
-        if (
-          property.type !== NodeTypes.DIRECTIVE
-          || property.name !== 'bind'
-          || property.rawName !== ':initial'
-        ) {
-          continue
-        }
-
-        // :initial should contain an object
-        if (
-          property.exp == null
-          || property.exp.ast == null
-          || typeof property.exp.ast === 'boolean'
-          || property.exp.ast.type !== 'ObjectExpression'
-        ) {
-          continue
-        }
-
-        // collect property and value
-        const initialState: Record<string, any> = {}
-        for (const p of property.exp.ast.properties) {
-          if (
-            p.type === 'ObjectProperty'
-            && p.key.type === 'Identifier'
-            && 'value' in p.value
-          ) {
-            initialState[p.key.name] = p.value.value
-          }
-        }
-
-        // bind parsed variant style to element
-        const normalized = stringifyStyle(initialState)
-        node.props.push({
-          type: NodeTypes.DIRECTIVE,
-          name: 'bind',
-          arg: createSimpleExpression('style', true, node.loc),
-          exp: createSimpleExpression(
-            JSON.stringify(normalized),
-            false,
-            node.loc,
-          ),
-          modifiers: [],
-          // loc: { ...node.loc, start: node.loc.end },
-          loc: node.loc,
-        })
-      }
-
-      return { props: [], needRuntime: true }
-    }
+    nuxt.options.vue.compilerOptions.directiveTransforms.motion
+      = motionDirectiveTransform
 
     // Transpile necessary packages
-    if (!nuxt.options.build.transpile)
-      nuxt.options.build.transpile = []
     const transpileList = [
       'defu',
       '@vueuse/motion',
       '@vueuse/shared',
       '@vueuse/core',
     ]
+    nuxt.options.build.transpile ??= []
     transpileList.forEach((pkgName) => {
       if (!nuxt.options.build.transpile.includes(pkgName))
         nuxt.options.build.transpile.push(pkgName)
